Rename steps in submit
